+++ src/cl.carousel/cl.carousel.js
@@ -117,31 +117,22 @@ var Cl = window.Cl || {};
 		},
 
 		next: function () {
-			// trigger event
-			this._fire('next');
-
 			// move
 			this.move(this.index + 1);
 
 			// trigger callback
-			this._fire('next', this);
+			this._fire('next');
 		},
 
 		previous: function () {
-			// trigger event
-			this._fire('previous');
-
 			// move
 			this.move(this.index - 1);
 
 			// trigger callback
-			this._fire('previous', this);
+			this._fire('previous');
 		},
 
 		move: function (index) {
-			// trigger event
-			this._fire('move');
-
 			// cancel if index is the same
 			if(index === this.index) return false;
 
@@ -180,13 +171,10 @@ var Cl = window.Cl || {};
 			this._accessibility();
 
 			// trigger callback
-			this._fire('move', this);
+			this._fire('move');
 		},
 
 		play: function () {
-			// trigger event
-			this._fire('play');
-
 			var that = this;
 			// start timer
 			this.timer = setInterval(function () {
@@ -197,13 +185,10 @@ var Cl = window.Cl || {};
 			this.autoplay = true;
 
 			// trigger event
-			this._fire('play', this);
+			this._fire('play');
 		},
 
 		stop: function () {
-			// trigger event
-			this._fire('stop');
-
 			// we just need to clear the intervall
 			clearInterval(this.timer);
 
@@ -211,24 +196,18 @@ var Cl = window.Cl || {};
 			this.autoplay = false;
 
 			// trigger event
-			this._fire('stop', this);
+			this._fire('stop');
 		},
 
 		update: function () {
-			// trigger event
-			this._fire('update');
-
 			// update gallery scripts
 			this.move(this.index);
 
 			// trigger event
-			this._fire('update', this);
+			this._fire('update');
 		},
 
 		destroy: function () {
-			// trigger event
-			this._fire('destroy');
-
 			this.viewport.removeAttr('style');
 			this.wrapper.removeAttr('style');
 			this.triggers.next.removeAttr('style');
@@ -243,7 +222,7 @@ var Cl = window.Cl || {};
 			this.stop();
 
 			// trigger event
-			this._fire('destroy', this);
+			this._fire('destroy');
 		},
 
 		_setIndex: function (index) {
@@ -293,16 +272,11 @@ var Cl = window.Cl || {};
 			}
 		},
 
-		_fire: function (keyword, scope) {
-			if(scope) {
-				// cancel if there is no callback found
-				if(this.callbacks[keyword] === undefined) return false;
-				// excecute callback
-				this.callbacks[keyword](scope);
-			} else {
-				// excecute event
-				$.event.trigger('carousel-' + keyword);
-			}
+		_fire: function (keyword) {
+			// cancel if there is no callback found
+			if(this.callbacks[keyword] === undefined) return false;
+			// excecute callback
+			this.callbacks[keyword](this);
 		}
 
 	});